Tests/LibWeb/Text/expected/UIEvents/middle-click-generated-content-link-autoscroll.txt

plain link scrollY: 0
generated-content link scrollY: 0
